backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
openai==1.3.5
google-generativeai==0.8.3
langchain==0.0.340
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
boto3==1.34.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
celery==5.3.4
httpx==0.25.1
tenacity==8.2.3
loguru==0.7.2
reportlab==4.0.4
authlib==1.3.0
alembic==1.12.1
psycopg2-binary==2.9.9
email-validator==2.1.0